[router] Unify forked functions with react-navigation
