+++ packages/expo-router/src/fork/__stubs__/window.tsx
