+++ packages/expo-router/src/fork/__tests__/useLinking.test.web.tsx
@@ -1,13 +1,24 @@
 import { act, render, waitFor } from '@testing-library/react-native';
 
-import { Screen } from '../../core/Screen';
-import { createNavigationContainerRef } from '../../core/createNavigationContainerRef';
-import { useNavigationBuilder } from '../../core/useNavigationBuilder';
-import { CommonActions, StackRouter } from '../../routers';
+import { RouterRegistryProvider } from '../../global-state/routerRegistry';
+import { Screen } from '../../react-navigation/core/Screen';
+import { createNavigationContainerRef } from '../../react-navigation/core/createNavigationContainerRef';
+import { useNavigationBuilder } from '../../react-navigation/core/useNavigationBuilder';
+import { CommonActions, StackRouter } from '../../react-navigation/routers';
 import { NavigationContainer } from '../NavigationContainer';
 import { createMemoryHistory } from '../createMemoryHistory';
 
 jest.mock('../createMemoryHistory');
+
+let mockNavigationRef: ReturnType<typeof createNavigationContainerRef>;
+
+jest.mock('../../global-state/storeContext', () => ({
+  useExpoRouterStore: () => ({
+    get state() {
+      return mockNavigationRef.current?.getRootState();
+    },
+  }),
+}));
 
 const history = {
   index: 0,
@@ -52,22 +63,25 @@ test('does not add browser history when preloading a stack route', async () => {
     );
   };
   const ref = createNavigationContainerRef<any>();
+  mockNavigationRef = ref;
   const onStateChange = jest.fn();
 
   render(
-    <NavigationContainer
-      ref={ref}
-      documentTitle={{ enabled: false }}
-      onStateChange={onStateChange}
-      linking={{
-        prefixes: [],
-        config: { screens: { home: 'home', details: 'details' } },
-      }}>
-      <Stack>
-        <Screen name="home" component={EmptyScreen} />
-        <Screen name="details" component={EmptyScreen} />
-      </Stack>
-    </NavigationContainer>
+    <RouterRegistryProvider>
+      <NavigationContainer
+        ref={ref}
+        documentTitle={{ enabled: false }}
+        onStateChange={onStateChange}
+        linking={{
+          prefixes: [],
+          config: { screens: { home: 'home', details: 'details' } },
+        }}>
+        <Stack>
+          <Screen name="home" component={EmptyScreen} />
+          <Screen name="details" component={EmptyScreen} />
+        </Stack>
+      </NavigationContainer>
+    </RouterRegistryProvider>
   );
 
   await waitFor(() => expect(ref.current).not.toBeNull());
