+++ packages/expo-router/src/fork/findFocusedRoute.tsx
@@ -1,8 +1,16 @@
-// Forked so we can access without importing any React Native code in Node.js environments.
-
+// Kept outside the vendored tree so Node.js code can import it without loading React Native.
 import type { InitialState } from '../react-navigation/routers';
 
-export function findFocusedRoute(state: InitialState) {
+type Result =
+  | {
+      key?: string;
+      name: string;
+      params?: object;
+      path?: string;
+    }
+  | undefined;
+
+export function findFocusedRoute(state: InitialState): Result {
   let current: InitialState | undefined = state;
 
   while (current?.routes[current.index ?? 0]?.state != null) {
@@ -10,5 +18,6 @@ export function findFocusedRoute(state: InitialState) {
   }
 
   const route = current?.routes[current?.index ?? 0];
+
   return route;
 }
