+++ packages/expo-router/src/fork/getPathFromState.ts
@@ -1,10 +1,12 @@
 import * as queryString from 'query-string';
 
+import { matchGroupName } from '../matchers';
 import { removeInternalExpoRouterParams } from '../navigationParams';
 import type { PathConfig, PathConfigMap } from '../react-navigation/native';
 import type { NavigationState, PartialState, Route } from '../react-navigation/routers';
 import * as expo from './getPathFromState-forks';
 import type { ExpoConfigItem, ExpoOptions } from './getPathFromState-forks';
+import { validatePathConfig } from './validatePathConfig';
 
 // START FORK
 export type Options<ParamList extends object> = ExpoOptions & {
@@ -87,10 +89,7 @@ export function getPathDataFromState<ParamList extends object>(
   }
 
   if (options) {
-    // START FORK
-    expo.validatePathConfig(options);
-    // validatePathConfig(options);
-    // END FORK
+    validatePathConfig(options);
   }
 
   // Create a normalized configs object which will be easier to use
@@ -118,6 +117,7 @@ export function getPathDataFromState<ParamList extends object>(
     let focusedParams: Record<string, any> | undefined;
     const focusedRoute = getActiveRoute(state);
     let currentOptions = configs;
+    const outerRouteIsGroup = matchGroupName(route.name) != null;
 
     // Keep all the route names that appeared during going deeper in config in case the pattern is resolved to undefined
     const nestedRouteNames: string[] = [];
@@ -135,6 +135,15 @@ export function getPathDataFromState<ParamList extends object>(
         // START FORK
         // This mutates allParams
         const currentParams = expo.fixCurrentParams(allParams, route, stringify);
+
+        if (
+          route.state === undefined &&
+          !outerRouteIsGroup &&
+          'screen' in route.params &&
+          typeof route.params.screen === 'string'
+        ) {
+          currentParams.screen = route.params.screen;
+        }
 
         // const currentParams = Object.fromEntries(
         //   Object.entries(route.params).map(([key, value]) => [
@@ -189,6 +198,11 @@ export function getPathDataFromState<ParamList extends object>(
         // route.push('/home/(a)')        --> This should navigate to /home/(a)/index
         // route.push('/home/(profile)')  --> This should navigate to /home/(profile)/me
         const screens = currentOptions[route.name]!.screens;
+
+        if (!outerRouteIsGroup) {
+          hasNext = false;
+          continue;
+        }
 
         // Determine what screen the user wants to navigate to. If no screen is specified, assume there is an index screen
         // In the examples above, this ensures that /home/(a) navigates to /home/(a)/index
