+++ packages/expo-router/src/fork/getStateFromPath.ts
@@ -2,11 +2,11 @@ import escape from 'escape-string-regexp';
 
 import { INTERNAL_SLOT_NAME } from '../constants';
 import type { PathConfigMap } from '../react-navigation/native';
-import { validatePathConfig } from '../react-navigation/native';
 import type { InitialState, NavigationState, PartialState } from '../react-navigation/routers';
 import { findFocusedRoute } from './findFocusedRoute';
 import type { ExpoOptions, ExpoRouteConfig } from './getStateFromPath-forks';
 import * as expo from './getStateFromPath-forks';
+import { validatePathConfig } from './validatePathConfig';
 
 export type Options<ParamList extends object> = ExpoOptions & {
   path?: string;
