+++ packages/expo-router/src/fork/useLinking.native.ts
@@ -63,18 +63,15 @@ export function useLinking(
     }
 
     if (enabled !== false && linkingHandlers.length) {
-      // TODO(@ubax): This check should be removed
-      if (linkingHandlers.length > 1) {
-        console.error(
-          [
-            'Looks like you have configured linking in multiple places. This is likely an error since deep links should only be handled in one place to avoid conflicts. Make sure that:',
-            "- You don't have multiple NavigationContainers in the app each with 'linking' enabled",
-            '- Only a single instance of the root component is rendered',
-          ]
-            .join('\n')
-            .trim()
-        );
-      }
+      console.error(
+        [
+          'Looks like you have configured linking in multiple places. This is likely an error since deep links should only be handled in one place to avoid conflicts. Make sure that:',
+          "- You don't have multiple NavigationContainers in the app each with 'linking' enabled",
+          '- Only a single instance of the root component is rendered',
+        ]
+          .join('\n')
+          .trim()
+      );
     }
 
     const handler = Symbol();
