+++ packages/expo-router/src/fork/validatePathConfig.ts
@@ -1,25 +1,71 @@
-const formatToList = (items: string[]) => items.map((key) => `- ${key}`).join('\n');
+const formatToList = (items: Record<string, string>) =>
+  Object.entries(items)
+    .map(([key, value]) => `- ${key} (${value})`)
+    .join('\n');
 
-export default function validatePathConfig(config: any, root = true) {
-  const validKeys = ['initialRouteName', 'screens', '_route'];
+const expoOptions = new Set(['preserveDynamicRoutes', 'preserveGroups', 'shouldEncodeURISegment']);
 
-  if (!root) {
-    validKeys.push('path', 'exact', 'stringify', 'parse');
-  }
+export function validatePathConfig(config: unknown, root = true) {
+  const validation = {
+    path: 'string',
+    initialRouteName: 'string',
+    screens: 'object',
+    ...(root
+      ? null
+      : {
+          _route: 'object',
+          exact: 'boolean',
+          stringify: 'object',
+          parse: 'object',
+        }),
+  };
 
-  const invalidKeys = Object.keys(config).filter((key) => !validKeys.includes(key));
-
-  if (invalidKeys.length) {
+  if (typeof config !== 'object' || config === null) {
     throw new Error(
-      `Found invalid properties in the configuration:\n${formatToList(
-        invalidKeys
-      )}\n\nDid you forget to specify them under a 'screens' property?\n\nYou can only specify the following properties:\n${formatToList(
-        validKeys
-      )}\n\nSee https://reactnavigation.org/docs/configuring-links for more details on how to specify a linking configuration.`
+      `Expected the configuration to be an object, but got ${JSON.stringify(config)}.`
     );
   }
 
-  if (config.screens) {
+  const validationErrors = Object.fromEntries(
+    Object.keys(config)
+      .filter((key) => !root || !expoOptions.has(key))
+      .map((key) => {
+        if (key in validation) {
+          const type = validation[key as keyof typeof validation];
+          // @ts-expect-error: we know the key exists
+          const value = config[key];
+
+          if (value !== undefined) {
+            if (typeof value !== type) {
+              return [key, `expected '${type}', got '${typeof value}'`];
+            }
+          }
+        } else {
+          return [key, 'extraneous'];
+        }
+
+        return null;
+      })
+      .filter(Boolean) as [string, string][]
+  );
+
+  if (Object.keys(validationErrors).length) {
+    throw new Error(
+      `Found invalid properties in the configuration:\n${formatToList(
+        validationErrors
+      )}\n\nYou can only specify the following properties:\n${formatToList(
+        validation
+      )}\n\nIf you want to specify configuration for screens, you need to specify them under a 'screens' property.\n\nSee https://reactnavigation.org/docs/configuring-links for more details on how to specify a linking configuration.`
+    );
+  }
+
+  if (root && 'path' in config && typeof config.path === 'string' && config.path.includes(':')) {
+    throw new Error(
+      `Found invalid path '${config.path}'. The 'path' in the top-level configuration cannot contain patterns for params.`
+    );
+  }
+
+  if ('screens' in config && config.screens) {
     Object.entries(config.screens).forEach(([_, value]) => {
       if (typeof value !== 'string') {
         validatePathConfig(value, false);
