+++ packages/expo-router/src/react-navigation/__tests__/exports.test.ts
@@ -2,10 +2,11 @@
 import * as Index from '../';
 
 describe('react-navigation/index re-exports', () => {
-  it('does not export static navigation APIs', () => {
+  it('does not export removed APIs', () => {
     expect(Index).not.toHaveProperty('createStaticNavigation');
     expect(Index).not.toHaveProperty('createComponentForStaticNavigation');
     expect(Index).not.toHaveProperty('createPathConfigForStaticNavigation');
+    expect(Index).not.toHaveProperty('NavigationContainer');
   });
 
   it.each([
