+++ packages/expo-router/src/react-navigation/core/__tests__/useStateForPath.test.ios.tsx
@@ -1,9 +1,9 @@
 import { act, render, screen } from '@testing-library/react-native';
 
+import { getPathFromState } from '../../../fork/getPathFromState';
 import type { ParamListBase } from '../../routers';
 import { Screen } from '../Screen';
 import { createNavigationContainerRef } from '../createNavigationContainerRef';
-import { getPathFromState } from '../getPathFromState';
 import { useNavigationBuilder } from '../useNavigationBuilder';
 import { useRoute } from '../useRoute';
 import { useStateForPath } from '../useStateForPath';
