+++ packages/expo-router/src/react-navigation/core/index.tsx
@@ -12,7 +12,7 @@ export { CurrentRenderContext } from './CurrentRenderContext';
 /**
  * @deprecated Will be removed in a future SDK.
  */
-export { findFocusedRoute } from './findFocusedRoute';
+export { findFocusedRoute } from '../../fork/findFocusedRoute';
 /**
  * @deprecated Will be removed in a future SDK.
  */
@@ -24,11 +24,11 @@ export { getFocusedRouteNameFromRoute } from './getFocusedRouteNameFromRoute';
 /**
  * @deprecated Will be removed in a future SDK.
  */
-export { getPathFromState } from './getPathFromState';
+export { getPathFromState } from '../../fork/getPathFromState';
 /**
  * @deprecated Will be removed in a future SDK.
  */
-export { getStateFromPath } from './getStateFromPath';
+export { getStateFromPath } from '../../fork/getStateFromPath';
 /**
  * @deprecated Will be removed in a future SDK.
  */
@@ -96,5 +96,5 @@ export { useStateForPath } from './useStateForPath';
 /**
  * @deprecated Will be removed in a future SDK.
  */
-export { validatePathConfig } from './validatePathConfig';
+export { validatePathConfig } from '../../fork/validatePathConfig';
 export * from '../routers';
