+++ packages/expo-router/src/react-navigation/native/__tests__/useLinkBuilder.test.ios.tsx
@@ -290,6 +290,7 @@ test('builds action from href in nested navigator', () => {
       payload: {
         name: 'Foo',
         params: {
+          id: '42',
           initial: true,
           screen: 'Bar',
           params: { id: '42' },
