+++ packages/expo-router/src/react-navigation/native/index.tsx
@@ -9,12 +9,6 @@ export { LinkingContext } from './LinkingContext';
  */
 export { LocaleDirContext } from './LocaleDirContext';
 /**
- * @deprecated `ExpoRoot` mounts the `NavigationContainer` automatically — there is no need
- * to render one manually in an Expo Router app. For tests, use `renderRouter` from
- * `expo-router/testing-library`. Will be removed in a future SDK.
- */
-export { NavigationContainer } from './NavigationContainer';
-/**
  * @deprecated Server-side rendering for Expo Router is handled by `@expo/server` and the
  * Expo Router static renderer. Will be removed in a future SDK.
  */
