--- packages/expo-router/src/react-navigation/native/useBackButton.tsx
@@ -1,6 +1,0 @@
-import type { NavigationContainerRef, ParamListBase } from '../core';
-
-export function useBackButton(_: React.RefObject<NavigationContainerRef<ParamListBase> | null>) {
-  // No-op
-  // BackHandler is not available on web
-}
