--- packages/expo-router/src/react-navigation/native/useDocumentTitle.native.tsx
@@ -1,3 +1,0 @@
-export function useDocumentTitle() {
-  // Noop for native platforms
-}
